split styles into files
