+++ src/components/chat-view.tsx
@@ -1,6 +1,6 @@
 import type { ChatMessage } from '@/lib/messages'
 import type { ReactElement } from 'react'
-import styles from '@/styles/app.module.css'
+import styles from '@/styles/chat-view.module.css'
 
 interface ChatViewProps {
     chat: ChatMessage[]
@@ -13,7 +13,7 @@ function ChatView(
         <section className={styles.chat}>
             {chat.map(({ role, content }, i) => (
                 <div className={styles.message} data-role={role} key={i}>
-                    <label>{role}</label>
+                    <label className="label">{role}</label>
                     <p>{content}</p>
                 </div>
             ))}
