+++ src/styles/app.module.css
@@ -4,58 +4,3 @@
     width: 100%;
     margin: auto;
 }
-
-.input {
-    position: relative;
-    display: grid;
-    grid-template-columns: 1fr 60px;
-    grid-template-rows: 25px;
-    margin: 5px;
-    border: 1px solid #fff;
-}
-
-.input > button {
-    background-color: #fff;
-    transition: background-color var(--timing);
-}
-
-.input > button:hover {
-    background-color: #f99;
-}
-
-.message {
-    display: flex;
-    flex-direction: column;
-    border: 1px solid var(--accent);
-    align-items: flex-start;
-    margin: 5px;
-}
-
-.message > label {
-    padding: 1px 5px;
-    background-color: var(--accent);
-}
-
-.message > p {
-    padding: 10px;
-}
-
-.message[data-role="user"] {
-    --accent: #f99;
-    align-items: flex-end;
-}
-
-.message[data-role="assistant"] {
-    --accent: #99f;
-}
-
-.message[data-role="system"] {
-    --accent: #9f9;
-}
-
-.input > button, .message > label {
-    color: #000;
-    font-size: 0.6rem;
-    font-weight: 500;
-    text-transform: uppercase;
-}
